# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatsApp-Community-Bot — Control API",
    "version": "1.0.0",
    "description": "Steuer- und Leseschnittstelle des Bots. Einziger vorgesehener Client ist die Android-App; ein direkter Zugriff auf Datenbank, WhatsApp-Session oder Serverdateien ist ausdruecklich nicht Teil dieser API.\n\n**Authentifizierung:** `Authorization: Bearer <accessToken>`. Access-Tokens leben 15 Minuten, Refresh-Tokens 30 Tage. Ein Refresh rotiert beide; ein bereits eingetauschter Refresh-Token beendet bei erneuter Vorlage die gesamte Geraete-Sitzung.\n\n**Fehler** haben immer dieselbe Form (siehe Schema `Error`) und tragen die `requestId`, die auch im Header `X-Request-Id` steht.\n\n**Rollen:** `viewer` liest, `admin` steuert zusaetzlich den Bot, `owner` verwaltet zusaetzlich Zugaenge. Das je Endpunkt noetige Recht steht als `x-required-permission`."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Anmelden und Token-Paar erhalten",
        "security": [],
        "description": "Solange **kein einziger** Zugang existiert, gilt einmalig `ACCESS_SECRET` als Passwort fuer den Benutzernamen `owner` und legt diesen Zugang an. Danach ist dieser Weg geschlossen.\n\nRate-Limit: 10 Anfragen je 15 Minuten und IP, geteilt mit `/auth/refresh`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "deviceLabel": {
                    "type": "string",
                    "maxLength": 64,
                    "description": "Anzeigename des Geraets in der Geraeteliste"
                  },
                  "appVersion": {
                    "type": "string",
                    "maxLength": 32
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — Zugang deaktiviert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Token-Paar erneuern",
        "security": [],
        "description": "Rotiert **beide** Tokens. Der eingereichte Refresh-Token wird dabei entwertet.\n\nWird ein bereits entwerteter Token erneut vorgelegt, gilt das als Diebstahl-Anzeichen: die komplette Geraete-Sitzung wird widerrufen und die Antwort ist `TOKEN_INVALID`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Erneuert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_INVALID oder TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Aktuelle Geraete-Sitzung beenden",
        "description": "Widerruft die Sitzung des benutzten Tokens samt aller ihrer Tokens.",
        "responses": {
          "200": {
            "description": "Abgemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Eigener Zugang und Rechte",
        "responses": {
          "200": {
            "description": "Zugangsdaten",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Eigene Geraete auflisten",
        "description": "Nur eigene Sitzungen. Tokens sind nie Teil der Antwort.",
        "responses": {
          "200": {
            "description": "Geraeteliste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions/{id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Ein Geraet abmelden",
        "description": "Nur eigene Sitzungen. Eine fremde Sitzungs-ID ergibt 404 — ihre Existenz wird nicht bestaetigt.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Sitzungs-ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Abgemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Zugaenge auflisten",
        "x-required-permission": "apiusers.manage",
        "responses": {
          "200": {
            "description": "Zugaenge",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiUser"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Zugang anlegen",
        "x-required-permission": "apiusers.manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "role"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 64,
                    "pattern": "^[a-z0-9._-]+$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 256
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "admin",
                      "owner"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Angelegt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — Name vergeben",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/users/{id}": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Zugang aktivieren oder deaktivieren",
        "x-required-permission": "apiusers.manage",
        "description": "Beim Deaktivieren werden alle Sitzungen dieses Zugangs sofort widerrufen. Der eigene Zugang laesst sich nicht deaktivieren.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Zugangs-ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "disabled"
                ],
                "properties": {
                  "disabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Geaendert",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — eigener Zugang",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Personen suchen und blaettern",
        "x-required-permission": "users.read",
        "description": "`limit` ist bei 100 gedeckelt; ein groesserer Wunsch wird heruntergesetzt, nicht abgelehnt. `q` braucht mindestens 2 Zeichen; `%` und `_` sind keine Platzhalter.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 120
            },
            "description": "Name, Nummer oder JID (Teilzeichenkette)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "activity",
                "xp",
                "warnings"
              ],
              "default": "name"
            }
          },
          {
            "name": "dir",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "groups",
                "warned",
                "muted"
              ],
              "default": "groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Treffer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "users",
                    "pagination"
                  ],
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserSummary"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{jid}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Eine Person im Detail",
        "x-required-permission": "users.read",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "WhatsApp-Kennung, z. B. 491701234567@s.whatsapp.net"
          }
        ],
        "responses": {
          "200": {
            "description": "Person",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access-Token aus /auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Maschinenlesbar und stabil — hierauf schalten, nicht auf message.",
                "examples": [
                  "UNAUTHENTICATED",
                  "TOKEN_EXPIRED",
                  "FORBIDDEN",
                  "VALIDATION_FAILED",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "BOT_NOT_READY"
                ]
              },
              "message": {
                "type": "string",
                "description": "Fuer Menschen. Darf sich aendern."
              },
              "requestId": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Auch im Header X-Request-Id."
              },
              "details": {
                "type": "array",
                "description": "Nur bei VALIDATION_FAILED.",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "problem": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "TokenPair": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "examples": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "integer",
            "description": "Sekunden bis der Access-Token ablaeuft."
          },
          "refreshExpiresIn": {
            "type": "integer"
          },
          "sessionId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/ApiUser"
          }
        }
      },
      "ApiUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "admin",
              "owner"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "disabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "Me": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/ApiUser"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deviceLabel": {
            "type": [
              "string",
              "null"
            ]
          },
          "appVersion": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "lastUsedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "revoked": {
            "type": "boolean"
          },
          "current": {
            "type": "boolean"
          }
        }
      },
...
      "UserSummary": {
        "type": "object",
        "properties": {
          "jid": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserIdentity"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "jid": {
                  "type": "string"
                },
                "name": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          },
          "xp": {
            "type": "integer"
          },
          "warnings": {
            "type": "integer"
          },
          "muted": {
            "type": "boolean"
          },
          "lastActive": {
            "type": [
              "integer",
              "null"
            ]
          },
          "inGroups": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}